Convert logbook into directive
