+++ ui-modules/home/app/index.js
@@ -32,10 +32,9 @@ import brooklynUserManagement from 'brooklyn-ui-utils/user-management/user-manag
 
 import mainState from 'views/main/main.controller';
 import mainDeployState from 'views/main/deploy/deploy.controller';
-import aboutState from 'views/about/about.controller.js';
-import logbookState from 'views/logbook/logbook.controller.js';
-
-import quickLaunchOverrides from '../../utils/providers/quick-launch-overrides.provider';
+import aboutState from 'views/about/about.controller';
+import brLogbook from 'brooklyn-ui-utils/logbook/logbook';
+import quickLaunchOverrides from 'brooklyn-ui-utils/providers/quick-launch-overrides.provider';
 
 import brandAngularJs from 'brand-angular-js';
 
@@ -43,7 +42,7 @@ const IS_PRODUCTION = process.env.NODE_ENV === 'production' || false;
 
 angular.module('brooklynHome', [ngAnimate, ngCookies, uiRouter, brCore, brServerStatus, brIconGenerator,
     brInterstitialSpinner, brooklynModuleLinks, brooklynUserManagement, brooklynQuickLaunch, mainState, mainDeployState,
-    aboutState, logbookState, quickLaunchOverrides, brandAngularJs])
+    aboutState, brLogbook, quickLaunchOverrides, brandAngularJs])
     .config(['$urlRouterProvider', '$logProvider', '$compileProvider', applicationConfig])
     .run(['$http', httpConfig]);
