+++ ui-modules/home/app/index.less
@@ -16,14 +16,17 @@
  * specific language governing permissions and limitations
  * under the License.
  */
+
+// Load first style rules
 @import '~brooklyn-shared/style/first.less';
 
+// Load util style rules
+@import '~brooklyn-ui-utils/logbook/logbook.less';
 @import '~brooklyn-ui-utils/yaml-editor/yaml-editor.less';
 @import '~brooklyn-ui-utils/quick-launch/quick-launch.less';
 
 // Add project less files here
 @import 'views/main/main';
-@import 'views/logbook/logbook';
 @import 'views/main/deploy/modal';
 @import 'components/onboarding/onboarding';
 
@@ -32,4 +35,4 @@
 }
 
 // Load last so that these style rules and var values trump others
-@import "~brooklyn-shared/style/last.less";
+@import '~brooklyn-shared/style/last.less';
