+++ ui-modules/home/app/views/about/about.controller.js
@@ -16,11 +16,12 @@
  * specific language governing permissions and limitations
  * under the License.
  */
+
 import angular from 'angular';
 import uiRouter from 'angular-ui-router';
-import serverApi from 'brooklyn-ui-utils/api/brooklyn/server.js';
+import serverApi from 'brooklyn-ui-utils/api/brooklyn/server';
 import {HIDE_INTERSTITIAL_SPINNER_EVENT} from 'brooklyn-ui-utils/interstitial-spinner/interstitial-spinner';
-import template from "./about.template.html";
+import template from './about.template.html';
 
 const MODULE_NAME = 'states.about';
 const BROOKLYN_VERSION = __BROOKLYN_VERSION__;
