+++ ui-modules/home/app/views/about/about.template.html
@@ -77,8 +77,6 @@
                 </tr>
                 </tbody>
             </table>
-            <h2>Logbook</h2>
-            <a class="btn btn-sm btn-default" href="#!/logbook"><i class="fa fa-eye"></i> Logbook</a>
         </div>
 
         <div class="col-md-6">
@@ -93,4 +91,7 @@
             </ul>
         </div>
     </div>
+
+    <h2>Logbook</h2>
+    <br-logbook></br-logbook>
 </div>
