--- ui-modules/home/app/views/logbook/logbook.template.html
@@ -1,73 +1,0 @@
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one
-  or more contributor license agreements.  See the NOTICE file
-  distributed with this work for additional information
-  regarding copyright ownership.  The ASF licenses this file
-  to you under the Apache License, Version 2.0 (the
-  "License"); you may not use this file except in compliance
-  with the License.  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing,
-  software distributed under the License is distributed on an
-  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  KIND, either express or implied.  See the License for the
-  specific language governing permissions and limitations
-  under the License.
--->
-<div class="container">
-    <div class="row">
-        <div class="col-md-12">
-            <h2>Logbook</h2>
-        </div>
-    </div>
-    <div class="row">
-        <div class="col-md-12">
-            <form>
-                <div>
-                    <div class="form-group col-md-3">
-                        <label for="initTime">Initial time</label>
-                        <input id="initTime" type="datetime-local" ng-model="initTime" class="form-control">
-                        <label for="finalTime">Final time</label>
-                        <input id="finalTime" type="datetime-local" ng-model="finalTime" class="form-control">
-
-                        <label for="numberOfItems">Number of lines</label>
-                        <input id="numberOfItems" type="number" ng-model="numberOfItems" class="form-control logbook" min="1">
-                        <label for="reverseOrder">Reverse order</label>
-                        <input id="reverseOrder" type="checkbox" ng-model="reverseOrder"><br/>
-                    </div>
-                    <div class="form-group col-md-9">
-                        <h4>Log level</h4>
-                        <label>
-                            <input type="checkbox" ng-model="allLevels" class=""/>All
-                        </label>
-                        <label ng-repeat="level in logLevels">
-                            <input type="checkbox" ng-model="level.selected" class=""/>{{level.name}}
-                        </label>
-                    </div>
-                    <div class="form-group col-md-9">
-                        <h4>Fields</h4>
-                        <label ng-repeat="field in logFields">
-                            <input type="checkbox"  ng-model="field.selected" class=""/>{{field.name}}
-                        </label>
-                    </div>
-
-                </div>
-                <div  class="form-group col-md-12">
-                    <button class="btn btn-outline btn-success" ng-click="vm.doQuery()" ng-disabled="waitingResponse">
-                        Query
-                    </button>
-                    <button class="btn btn-outline" ng-click="vm.resetForm()">
-                        Reset form
-                    </button>
-                </div>
-            </form>
-        </div>
-        <div>
-            <textarea id="logContent" class="logbook-text" ng-model="results"></textarea>
-        </div>
-        </form>
-    </div>
-</div>
-</div>
